Fix rollout grid. Add tests for controller management.

+++ hawkbit-ui/src/main/java/org/eclipse/hawkbit/ui/rollout/rollout/ProxyRollout.java
@@ -30,7 +30,7 @@ public class ProxyRollout {
 
     private String modifiedDate;
 
-    private Long numberOfGroups;
+    private Integer numberOfGroups;
 
     private Boolean isActionRecieved = Boolean.FALSE;
 
@@ -149,7 +149,7 @@ public class ProxyRollout {
     /**
      * @return the numberOfGroups
      */
-    public Long getNumberOfGroups() {
+    public Integer getNumberOfGroups() {
         return numberOfGroups;
     }
 
@@ -157,7 +157,7 @@ public class ProxyRollout {
      * @param numberOfGroups
      *            the numberOfGroups to set
      */
-    public void setNumberOfGroups(final Long numberOfGroups) {
+    public void setNumberOfGroups(final Integer numberOfGroups) {
         this.numberOfGroups = numberOfGroups;
     }
